+++ ergotree-ir/src/serialization/sigma_byte_reader.rs
@@ -51,7 +51,7 @@ pub fn from_bytes<T: AsRef<[u8]>>(bytes: T) -> SigmaByteReader<PeekableReader<Cu
 }
 
 /// Sigma byte reader trait with a constant store to resolve segregated constants
-pub trait SigmaByteRead: ReadSigmaVlqExt {
+pub trait SigmaByteRead: ReadSigmaVlqExt + Peekable {
     /// Constant store with constants to resolve constant placeholder types
     fn constant_store(&mut self) -> &mut ConstantStore;
 
@@ -65,7 +65,7 @@ pub trait SigmaByteRead: ReadSigmaVlqExt {
     fn val_def_type_store(&mut self) -> &mut ValDefTypeStore;
 }
 
-impl<R: Peekable> Read for SigmaByteReader<R> {
+impl<R: Read> Read for SigmaByteReader<R> {
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
         self.inner.read(buf)
     }
@@ -77,7 +77,7 @@ impl<R: Peekable> Peekable for SigmaByteReader<R> {
     }
 }
 
-impl<R: ReadSigmaVlqExt> SigmaByteRead for SigmaByteReader<R> {
+impl<R: ReadSigmaVlqExt + Peekable> SigmaByteRead for SigmaByteReader<R> {
     fn constant_store(&mut self) -> &mut ConstantStore {
         &mut self.constant_store
     }