+++ ergotree-ir/src/serialization/types.rs
@@ -376,7 +376,6 @@ mod tests {
 
         #[test]
         fn ser_roundtrip(v in any::<SType>()) {
-            dbg!(&v);
             prop_assert_eq![sigma_serialize_roundtrip(&v), v];
         }
     }